Build Cenedril + the Windows VST3/Vizia foundation (M0+M1)
Stand up the first Windows-only new VST (Cenedril, the passthrough Visualizer)
and the shared Windows build + editor foundation the other new VSTs (Calóma,
Lúmedir) will reuse.

Shared Windows foundation:
- xtask Windows `.vst3` bundle path: cross-build with cargo-xwin (MSVC ABI) via
  a `cargo xwin build` arm in `build_release`, a `Contents/x86_64-win/` bundle
  layout (`windows_bundle.rs`), and a `make build-windows` target parameterized
  by WINDOWS_PLUGINS (incl. an idempotent lld-link case-fix for the xwin SDK).
- lindelion-plugin-shell now compiles for `x86_64-pc-windows-msvc`: normalize
  the vst3 COM flag/controller constants that are i32 on Windows / u32 elsewhere.
- Shared editor enabler: `IPlugView` `kPlatformTypeHWND` support in vst3_view.
- lindelion-ui gains a `cfg(windows)` vizia + raw-window-handle dep section; the
  Vizia/baseview/skia stack cross-compiles for Windows.

Cenedril (plugins/cenedril):
- Effect crate with a bit-exact, zero-latency stereo passthrough processor and
  the VST3 processor/controller/factory scaffold; CENEDRIL_VST3_BUNDLE_METADATA.
- M1 editor: a Windows `IPlugView`→`HWND` Vizia/baseview attach with a static
  placeholder view (cenedril_vizia), gated like the macOS editors.

Docs: ADR-0023 reframed to Windows + Vizia (correcting the earlier egui framing
and the false "Vizia builds on Linux in make ci" claim — vizia is target-gated,
so the Vizia editor is verified by the Windows build, not make ci); plan + step
files for Cenedril M0/M1.

make ci is green on Linux; `make build-windows` stages a working Cenedril.vst3.
Windows-host runtime validation (load/render/resize) is left for manual/Galad.

+++ crates/lindelion-plugin-shell/src/vst3_component.rs
@@ -221,6 +221,10 @@ pub struct Vst3BusInfo {
     pub flags: u32,
 }
 
+// The `vst3` COM bindings type bus/component flag constants as `u32` on Linux/macOS but `i32` on
+// the Windows MSVC target; normalize with `as u32` (identity for these small flag values). The
+// allow keeps `make ci`'s clippy green on Linux, where the cast is a no-op.
+#[allow(clippy::unnecessary_cast)]
 impl Vst3BusInfo {
     pub const fn audio_input(channel_count: i32, name: &'static str) -> Self {
         Self {
@@ -229,7 +233,7 @@ impl Vst3BusInfo {
             channel_count,
             name,
             bus_type: BusTypes_::kMain,
-            flags: BusInfo_::BusFlags_::kDefaultActive,
+            flags: BusInfo_::BusFlags_::kDefaultActive as u32,
         }
     }
 
@@ -251,7 +255,7 @@ impl Vst3BusInfo {
             channel_count,
             name,
             bus_type: BusTypes_::kMain,
-            flags: BusInfo_::BusFlags_::kDefaultActive,
+            flags: BusInfo_::BusFlags_::kDefaultActive as u32,
         }
     }
 
@@ -262,7 +266,7 @@ impl Vst3BusInfo {
             channel_count,
             name,
             bus_type: BusTypes_::kMain,
-            flags: BusInfo_::BusFlags_::kDefaultActive,
+            flags: BusInfo_::BusFlags_::kDefaultActive as u32,
         }
     }
 
@@ -273,7 +277,7 @@ impl Vst3BusInfo {
             channel_count,
             name,
             bus_type: BusTypes_::kMain,
-            flags: BusInfo_::BusFlags_::kDefaultActive,
+            flags: BusInfo_::BusFlags_::kDefaultActive as u32,
         }
     }
 }
